+++ src/main/java/atom/marvelores/item/ModItemGroups.java
@@ -25,6 +25,8 @@ public class ModItemGroups {
                         entries.add(ModItems.VIBRANIUM_AXE);
                         entries.add(ModItems.VIBRANIUM_HOE);
 
+                        /*entries.add(ModItems.VIBRANIUM_SHIELD);*/
+
                         entries.add(ModItems.VIBRANIUM_HELMET);
                         entries.add(ModItems.VIBRANIUM_CHESTPLATE);
                         entries.add(ModItems.VIBRANIUM_LEGGINGS);